endurance_report.py: fix disproportional graph visualization when using browser zoom
The graphs in the endurance reports are generated inside one <table>
with embedded <table> elements. For one of these outer level <table>s,
the embedded <table> widths are not equal.

This gives problems when zooming with web browser, because the browser
then alters the widths with different coefficient, leading to misleading
visualization.

Fixes: NB#242899 - Endurance report memory usage graph pixel width not
respected by browser zoom, results to misleading visualization