save-incremental-endurance-stats: collect `df' output later
The save-incremental-endurance-stats execution has been seen to block at
the `df' execution when the system has been extremely loaded. In this
case we manage to collect only a partial snapshot. Move the `df'
invocation to later position in the script, so that we may be able to
collect as much information as possible before invoking it.