endurance_plot: fix incorrect entries in some histogram plots due to gnuplot data sharing
We were using the same input data file (names) to generate different
graphs, but we want the data to be slightly different for line and
histogram graphs with respect to zero entries. This was causing
incorrect entries in some of the histogram graphs with processes that
are not present from the start of the endurance data.

To fix this, generate unique data files for each graph, so that no
gnuplot input data sharing is used.