endurance_plot: also fill `name_by_pid' from usage.csv
The `name_by_pid' hash is populated from the smaps data, but sometimes
we may have some short living processes that are present in the
usage.csv file, but not in the smaps data. In these cases it is useful
to take the process names from the usage.csv file in addition to the
smaps data.