endurance_plot: allow zeropad removal for linegraphs
We generate both linegraphs and histograms based on some of the data we
parse. For the histograms, leading zeroes are required for all entries
to get correct gnuplot results, but for the linegraphs we often do not
want the leading zeroes. Extend the emitting logic to allow for the zero
padding to be removed for the benefit of linegraphs.