post processing modularization
Split the original script into several smaller Perl modules, that are
easier to understand and to test. The perl modules may also be used
outside of sp-endurance, but we do not provide any API stability
guarantees.

The parsing code is greatly simplified, and the parsing step now simply
parses data to memory, which is then consumed by the various functions
that create the graphs. Note that in order to save memory, the parser
module only stores in memory the data that we will use.

There is now an option for improving the parsing performance that is
automatically enabled if the Perl "Inline" module from CPAN is
available. The '/proc/pid/smaps' parsing has been implemented in C,
allowing for improved performance.

As a side effect of all these changes, the endurance_plot script had to
be almost completely rewritten.