Use common `endurance-' prefix for executables
Rename the executables to share a common `endurance-' prefix for
improved clarity and easier discovery. The data collection script
`save-incremental-endurance-stats' is renamed to `endurance-snapshot',
and the word "measurement" is replaced with "snapshot" it other places.

Compatibility with the old names can be retained by creating symbolic
links at install time. A new Makefile target is added for this purpose.