measure/endurance-snapshot: proc2csv, sp-noncached from script directory
Check if proc2csv and sp-noncached are located in the same directory as
the endurance-snapshot script, so that the collection script and the
executables can be e.g. in custom location outside PATH.

While at it, make sp-noncached optional, just use plain and simple 'cat'
if it is not available.