measure/endurance-snapshot: simpler smaps.cap capture unless sp_smaps_snapshot is available
Simplify collection of /proc/*/smaps (or /proc/*/maps) when
sp_smaps_snapshot is not available.

Idea is to avoid exec'ing awk & cat for each process in system.

This also means that /proc/*/status data is not included in smaps.cap,
but it's collected by proc2csv, so we do not lose any information from
sp-endurance point-of-view.