postproc-lib: parse 'name' out of /proc/pid/stat data
With the simplified smaps.cap format introduced in commit
295bcaa687f0a75c3c0160f653fe3f32feb04f68 ("measure/endurance-snapshot:
simpler smaps.cap capture unless sp_smaps_snapshot is available"), we
could be missing names for kthreads in the graphs, as the names are no
longer stored in the smaps.cap data file.

To fix, check if we have the name in the /proc/pid/stat data.