endurance-snapshot: collect files to temporary directory
Collect files to temporary directory (append shell PID to snapshot
directory name), and move/rename that directory to the final path when
finished.

This can be useful in some cases, e.g. when manipulating collected data
in the same environment where we're collecting, so that we can clearly
see which snapshot directories are finished.