proc2csv: store up to 2048 bytes of process cmdline
Take more bytes of process cmdlines, to try to ensure we have more
complete information available, as we utilize these now in the
endurance-plot graphs.

Allocate the strings with strdup() to avoid always using 2048 bytes per
process.

Note that currently HAVE_VALGRIND_H is not automatically defined
anywhere, so it needs to be manually set if wanted.